refactor: introduce CaseChat provider
